// src/model/Response/Hotel/RoomGetByHotelIDRes.tsx
export interface RoomGetByHotelIDRes {
  HRID: number;
  hotel_ID: number;
  name: string;
  price: number;
  Number_of_guests: number;
  Number_of_rooms: number;
  room_type_ID: number;
  type_room: string;
  room_view_type_ID: number;
  type_view_name_room: string;
  room_status_ID: number;
  status_name_room: string;
}